Run rustfmt and disable clippy in Travis (#39)
Builds that fail for issues unrelated to pull requests are useless. This
PR gets the build passing again and hopefully it will stay that way.

One caveat is that clippy is now disabled in Travis builds. It is giving
me trouble running locally at the moment; I think the best approach is
to keep it disabled until it is incorporated into rustup.