Update GenericArray to 0.11.1
This is the newest version; we were depending on an ancient version
before.

I am not convinced that this is enough to fix dimensioned for `no_std`
environments and would like to add a test that enforces it working on
`no_std`.

Also includes some formatting and the addition of the `std` default
feature.